aula24/requirements.txt
fastapi
uvicorn
pyjwt
httpx
requests
